+++ examples/grid.py
@@ -1,15 +1,13 @@
 from textual.app import App
 from textual import events
-from textual.view import View
 from textual.widgets import Placeholder
-from textual.layouts.grid import GridLayout
 
 
 class GridTest(App):
     async def on_load(self, event: events.Load) -> None:
         await self.bind("q,ctrl+c", "quit", "Quit")
 
-    async def on_startup(self, event: events.Startup) -> None:
+    async def on_mount(self, event: events.Mount) -> None:
 
         grid = await self.view.dock_grid()