+++ src/textual/views/__init__.py
@@ -1,1 +1,2 @@
 from ._dock_view import DockView, Dock, DockEdge
+from ._grid_view import GridView