+++ src/textual/widgets/__init__.py
@@ -1,8 +1,16 @@
 from ._footer import Footer
 from ._header import Header
-from ._button import Button
+from ._button import Button, ButtonPressed
 from ._placeholder import Placeholder
 from ._scroll_view import ScrollView
 from ._static import Static
 
-__all__ = ["Footer", "Header", "Button", "Placeholder", "ScrollView", "Static"]
+__all__ = [
+    "Footer",
+    "Header",
+    "Button",
+    "ButtonPressed",
+    "Placeholder",
+    "ScrollView",
+    "Static",
+]